kodo: add layer to use systematic for last needed symbol
If one is using status vectors to update decoder status at the encoder,
situations might occur, where the encoder tries to encoder over a single
symbol. (E.g. when encoder knows that the decoder was partial complete
and only one new symbol is added.) In this case, the encoder should use
systematic encoding, as it has no other packets to include.

This patch adds a layer to intersect the encoding process, detect
when the above situation occurs, and then use systematic encoding.

The layer relies on the two previous added layers to detect when
systematic should be used: encoder_status_vector
and storage_status vector.

Signed-off-to: SteinWurf ApS <contact@steinwurf.com>

+++ src/kodo/storage_status_vector.hpp
@@ -25,7 +25,7 @@ namespace kodo
 
     protected:
 
-        typedef uint32_t status_block;
+        typedef uint8_t status_block;
         typedef boost::dynamic_bitset<status_block> status_vector;
 
     public:
